admin login and layout page implemented

+++ app/views/admin_users/index.blade.php
@@ -1,10 +1,11 @@
+
 @extends('layouts.scaffold')
 
 @section('main')
 
-<h1>All Admin_users</h1>
+<h1>All Admin Users</h1>
 
-<p>{{ link_to_route('admin_users.create', 'Add New Admin_user', null, array('class' => 'btn btn-lg btn-success')) }}</p>
+{{--<p>{{ link_to_route('admin_users.create', 'Add New Admin_user', null, array('class' => 'btn btn-lg btn-success')) }}</p>--}}
 
 @if ($admin_users->count())
 	<table class="table table-striped">
@@ -13,7 +14,7 @@
 				<th>Name</th>
 				<th>Email</th>
 				<th>Password</th>
-				<th>Is_superuser</th>
+				<th>Superuser</th>
 				<th>&nbsp;</th>
 			</tr>
 		</thead>
@@ -25,16 +26,26 @@
 					<td>{{{ $admin_user->email }}}</td>
 					<td>{{{ $admin_user->password }}}</td>
 					<td>{{{ $admin_user->is_superuser }}}</td>
-                    <td>
-                        {{ Form::open(array('style' => 'display: inline-block;', 'method' => 'DELETE', 'route' => array('admin_users.destroy', $admin_user->id))) }}
-                            {{ Form::submit('Delete', array('class' => 'btn btn-danger')) }}
-                        {{ Form::close() }}
-                        {{ link_to_route('admin_users.edit', 'Edit', array($admin_user->id), array('class' => 'btn btn-info')) }}
-                    </td>
+
+                    @if (Auth::user()->is_superuser)
+                        <td>
+
+                            {{ Form::open(array('style' => 'display: inline-block;', 'method' => 'DELETE', 'route' => array('admin_users.destroy', $admin_user->id), 'onsubmit' => 'return ConfirmDelete()')) }}
+                                {{ Form::submit('Delete', array('class' => 'btn btn-danger')) }}
+                            {{ Form::close() }}
+                        </td>
+                        <td>
+                            {{ link_to_route('admin_users.edit', 'Edit', array($admin_user->id), array('class' => 'btn btn-info')) }}
+                        </td>
+                    @endif
 				</tr>
 			@endforeach
 		</tbody>
 	</table>
+
+    @if (Auth::user()->is_superuser)
+        <p>{{ link_to_route('admin_users.create', 'Add Admin', null, array('class' => 'btn btn-lg btn-success pos_right')) }}</p>
+    @endif
 @else
 	There are no admin_users
 @endif
